pyproject.toml
[project]
name = "logo-stamper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.97.1",
    "pillow>=11.3.0",
]
